Implemented advanced navigation system using GoRouter, replacing traditional Navigator. Centralized route configuration in app_router.dart, updated main application to use MaterialApp.router, and enhanced home page with modern UI and navigation buttons. Updated dependencies and documentation to reflect changes.
