+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   flutter_svg: ^2.2.0
+  go_router: ^16.0.0
 
 dev_dependencies:
   flutter_test:
